New check: Sys.Break is always caught explicitly
The `Sys.Break` exception is a bit special in that it is raised automatically by
the runtime. Each function application (and maybe other kind of constructs) can
raise it. So, for example it is in general an error to do:

```ocaml
try Some (f x) with
| e -> None
```

If `Sys.Break` is caught it can alter the result and the program will go on
"normally". Usually it is a good idea to re-raise until a proper handler is
found.

```ocaml
try Some (f x) with
| Sys.Break as e -> raise e
| e -> None
```

This new check flags `try` constructs that have both:

  - no `Sys.Break` branch
  - a wildcard branch