Allow "match () with ..." as an exception
Rationale:

The following piece of code:

```ocaml
match () with
| _ when guard_a -> result_a
| _ when guard_b -> result_b
| _ -> default
```

is equivalent to:

```ocaml
if guard_a then
  result_a
else if guard_b then
  result_b
else
  default
```

But makes it possible to reorder the guards.

There is only one constructor for `unit`, so the only meaningful use of `match
() with` is to use guards.