Add a plugin system to configure rules
This adds a mechanism to configure ocamllint.

A configuration object determines whether a warning should be emitted or not
(the default one always emits a warning).

The ppx module can take a compiled object file as an argument. In that case, it
will dynlink it and execute its initialization command.

That command can be `set_config` which will override the default configuration.