Only run when a compiler is running.
If another tool (such as `ocamldep`) is running, do not run (or rather, run
`default_mapper`).

There is a small caveat in the implementation: `tool_name ()` will return a
meaningful value only from the mapper itself. So, it is not possible to call it
before the `register` call, because there it will return `"_none_"`.

Closes #12